whisper-srt/diarize/requirements.txt
# Speaker diarization + whisper transcription (merge to SRT with speaker labels)
# Install: pip install -r requirements.txt
# Requires: Hugging Face token with access to pyannote/speaker-diarization-3.1
pyannote.audio>=3.0.0
requests>=2.28.0
